+++ instances/src/main/scala/pl/iterators/kebs/instances/time/ZoneOffsetString.scala
@@ -1,0 +1,14 @@
+package pl.iterators.kebs.instances.time
+
+import pl.iterators.kebs.instances.InstanceConverter
+import pl.iterators.kebs.instances.time.ZoneOffsetString.ZoneOffsetFormat
+
+import java.time.ZoneOffset
+
+trait ZoneOffsetString {
+  implicit val zoneOffsetFormatter: InstanceConverter[ZoneOffset, String] =
+    InstanceConverter[ZoneOffset, String](_.toString, ZoneOffset.of, Some(ZoneOffsetFormat))
+}
+object ZoneOffsetString {
+  private[instances] val ZoneOffsetFormat = "ISO-8601 standard format e.g. +01:00"
+}