+++ instances/src/main/scala/pl/iterators/kebs/instances/time/mixins/InstantEpochMilliLong.scala
@@ -1,0 +1,10 @@
+package pl.iterators.kebs.instances.time.mixins
+
+import pl.iterators.kebs.instances.InstanceConverter
+
+import java.time.Instant
+
+trait InstantEpochMilliLong {
+  implicit val instantEpochMilliFormatter: InstanceConverter[Instant, Long] =
+    InstanceConverter.apply[Instant, Long](_.toEpochMilli, Instant.ofEpochMilli)
+}